feat: refactor utility imports and add getTextContent function for improved text extraction
