+++ src/Components/UI/pagination.tsx
@@ -5,8 +5,8 @@ import {
   MoreHorizontalIcon,
 } from "lucide-react"
 
-import { cn } from "@/lib/utils"
 import { Button, buttonVariants } from "@/Components/UI/button"
+import { cn } from "@/lib/util"
 
 function Pagination({ className, ...props }: React.ComponentProps<"nav">) {
   return (
