+++ src/lib/getGetTextContent.ts
